+++ windows/window_manager_plugin.cpp
@@ -108,6 +108,10 @@ std::optional<LRESULT> WindowManagerPlugin::HandleWindowProc(HWND hWnd,
                                                              LPARAM lParam) {
   std::optional<LRESULT> result = std::nullopt;
 
+  if (message == WM_DPICHANGED) {
+    window_manager->pixel_ratio_ = (float) LOWORD(wParam) / USER_DEFAULT_SCREEN_DPI;
+  }
+
   if (message == WM_NCCALCSIZE) {
     // This must always be first or else the one of other two ifs will execute
     //  when window is in full screen and we don't want that
@@ -163,13 +167,21 @@ std::optional<LRESULT> WindowManagerPlugin::HandleWindowProc(HWND hWnd,
     MINMAXINFO* info = reinterpret_cast<MINMAXINFO*>(lParam);
     // For the special "unconstrained" values, leave the defaults.
     if (window_manager->minimum_size_.x != 0)
-      info->ptMinTrackSize.x = window_manager->minimum_size_.x;
+      info->ptMinTrackSize.x =
+          static_cast<LONG> (window_manager->minimum_size_.x *
+          window_manager->pixel_ratio_);
     if (window_manager->minimum_size_.y != 0)
-      info->ptMinTrackSize.y = window_manager->minimum_size_.y;
+      info->ptMinTrackSize.y =
+          static_cast<LONG> (window_manager->minimum_size_.y *
+          window_manager->pixel_ratio_);
     if (window_manager->maximum_size_.x != -1)
-      info->ptMaxTrackSize.x = window_manager->maximum_size_.x;
+      info->ptMaxTrackSize.x =
+          static_cast<LONG> (window_manager->maximum_size_.x *
+          window_manager->pixel_ratio_);
     if (window_manager->maximum_size_.y != -1)
-      info->ptMaxTrackSize.y = window_manager->maximum_size_.y;
+      info->ptMaxTrackSize.y =
+          static_cast<LONG> (window_manager->maximum_size_.y *
+          window_manager->pixel_ratio_);
     result = 0;
   } else if (message == WM_NCACTIVATE) {
     if (wParam == TRUE) {
@@ -295,7 +307,15 @@ std::optional<LRESULT> WindowManagerPlugin::HandleWindowProc(HWND hWnd,
     } else {
       _EmitEvent("hide");
     }
+  } else if (message == WM_WINDOWPOSCHANGED) {
+    if (window_manager->IsAlwaysOnBottom()) {
+        const flutter::EncodableMap& args = {
+		  {flutter::EncodableValue("isAlwaysOnBottom"),
+            		   flutter::EncodableValue(true)}};
+	    window_manager->SetAlwaysOnBottom(args);
+	  }
   }
+  
   return result;
 }
 
@@ -444,6 +464,14 @@ void WindowManagerPlugin::HandleMethodCall(
         std::get<flutter::EncodableMap>(*method_call.arguments());
     window_manager->SetAlwaysOnTop(args);
     result->Success(flutter::EncodableValue(true));
+  } else if (method_name.compare("isAlwaysOnBottom") == 0) {
+    bool value = window_manager->IsAlwaysOnBottom();
+    result->Success(flutter::EncodableValue(value));
+  } else if (method_name.compare("setAlwaysOnBottom") == 0) {
+    const flutter::EncodableMap& args =
+        std::get<flutter::EncodableMap>(*method_call.arguments());
+    window_manager->SetAlwaysOnBottom(args);
+    result->Success(flutter::EncodableValue(true));
   } else if (method_name.compare("getTitle") == 0) {
     std::string value = window_manager->GetTitle();
     result->Success(flutter::EncodableValue(value));
@@ -520,7 +548,7 @@ void WindowManagerPlugin::HandleMethodCall(
   } else {
     result->NotImplemented();
   }
-}
+ }
 
 }  // namespace